+++ prop/example/tautology.test.ts
