+++ prop/example/tautology.ts
